#12 - moving functions to functionslib part2
